Cogs

+++ cogs/CoC.py
@@ -2,7 +2,7 @@ from discord.ext import commands
 import discord
 from main import *
 import mariadb
-from time import time,sleep
+from time import time, sleep
 import asyncio
 import aiohttp
 from main import config
@@ -14,8 +14,8 @@ class CocDB:
         self.COCDB = config.config['db']
         self.FETCHPLAYER_API_URL = f'https://api.clashofclans.com/v1/players/'
         self.db = mariadb.connect(
-            user=config.config['user'], 
-            host=config.config['host'], 
+            user=config.config['user'],
+            host=config.config['host'],
             password=config.config['password'],
             port=config.config['port'])
         self.c = self.db.cursor(buffered=True)
@@ -30,7 +30,7 @@ cocdb = CocDB()
 
 
 class CocRaw:
-    
+
     def __init__(self):
         self.DATA = {}
         self.LEGEND_SEASONS = [
@@ -41,27 +41,27 @@ class CocRaw:
             '2019-01', '2019-02', '2019-03', '2019-04', '2019-05', '2019-06', '2019-07', '2019-08', '2019-09', '2019-10', '2019-11', '2019-12',
             '2020-01', '2020-02', '2020-03', '2020-04', '2020-05', '2020-06', '2020-07', '2020-08', '2020-09', '2020-10', '2020-11', '2020-12',
             '2021-01', '2021-02', '2021-03', '2021-04', '2021-05', '2021-06', '2021-07'
-            ]  # ,'2021-08','2021-09','2021-10','2021-11','2021-12']
+        ]  # ,'2021-08','2021-09','2021-10','2021-11','2021-12']
         print('Preloading datasets...')
         for season in self.LEGEND_SEASONS:
-            sleep(1)
             cocdb.c.execute('SELECT * FROM `{}`'.format(season))
             result = cocdb.c.fetchall()
             self.DATA[season] = {}
             for record in result:
                 self.DATA[season][record[0]] = {
-                                    'season': season,
-                                    'name': record[1],
-                                    'expLevel': record[2],
-                                    'trophies': record[3],
-                                    'attackWins': record[4],
-                                    'defenseWins': record[5],
-                                    'rank': record[6],
-                                    'clantag': record[7],
-                                    'clanname': record[8],
-                                    'clanbadgeUrl': record[9]}
-        
-        
+                    'season': season,
+                    'tag': record[0],
+                    'name': record[1],
+                    'expLevel': record[2],
+                    'trophies': record[3],
+                    'attackWins': record[4],
+                    'defenseWins': record[5],
+                    'rank': record[6],
+                    'clantag': record[7],
+                    'clanname': record[8],
+                    'clanbadgeUrl': record[9]}
+
+
 cocraw = CocRaw()
 cocdb.c.close()
 cocdb.db.close()
@@ -71,6 +71,7 @@ class CoC(commands.Cog):
 
     def __init__(self, bot):
         self.bot = bot
+        self.COC_CLIENT_MEMBER_LAST_ACTIVITY_CACHE = {}  # user last seen
 
     async def searchtagbyseason(self, season, tag):
         try:
@@ -81,35 +82,39 @@ class CoC(commands.Cog):
     @commands.command()
     @commands.cooldown(2, 15, commands.BucketType.guild)
     @commands.guild_only()
-    async def tag(self, ctx, tag):
+    async def tag(self, ctx, tag: str):
+        if not ctx.author.id in config.config['whitelist']:
+            return await ctx.send(embed=embeds.Error._text_to_embed(self.bot, ctx, 'You\'re not allowed to use this command.'))
         try:
             tag = '#' + tag.replace('#', '')
             async with aiohttp.ClientSession() as session:
-                async with session.get(f'https://api.clashofclans.com/v1/players/{tag.replace("#","%23")}',headers={"Accept": "application/json", "authorization": f"Bearer {config.config['key']}"}) as request:
+                async with session.get(f'https://api.clashofclans.com/v1/players/{tag.replace("#","%23")}', headers={"Accept": "application/json", "authorization": f"Bearer {config.config['key']}"}) as request:
                     if not (request.status == 200):
                         print(request)
                         return await ctx.send(embed=embeds.Error._text_to_embed(self.bot, ctx, 'This player doesn\'t exist.'))
                     r = await request.json()
                     playerName = r['name']
             s = time()
-            msg = await ctx.send(embed=embeds.Loading._text_to_embed(self.bot, ctx, f'Looping through 20M+ user records for the player **{playerName}**'))
+            msg = await ctx.send(embed=embeds.Loading._text_to_embed(self.bot, ctx, f'Searching through 25m+ user records for the player **{playerName}**'))
             async with ctx.typing():
-                results =  [await self.searchtagbyseason(season, tag) for season in cocraw.LEGEND_SEASONS]
+                results = [await self.searchtagbyseason(season, tag) for season in cocraw.LEGEND_SEASONS]
                 e = time()
                 results = [result for result in results if result != None]
                 if not results:
                     return await ctx.send(embed=embeds.Error._text_to_embed(self.bot, ctx, f'User **{playerName}** not found in our database. Note: User must\'ve been in the legend league once.'))
                 embed = discord.Embed(description=', '.join(
                     '`' + x['season'] + '`' for x in results) + '\n\n✅ **Show all results**\n❌ **Delete this embed**', color=DEFAULT_COLOR, timestamp=ctx.message.created_at)
-                embed.set_footer(text=f'Took {round(e-s,3)}s | Requested by n?gga {ctx.author.name}')
-                embed.set_author(name=f'Seasons of the player {playerName} [{tag}]',icon_url=ctx.author.avatar_url)
+                embed.set_footer(
+                    text=f'Took {round(e-s,3)}s | Requested by {ctx.author.name}')
+                embed.set_author(
+                    name=f'Seasons of the player {playerName} [{tag}]', icon_url=ctx.author.avatar_url)
                 await msg.edit(embed=embed)
                 await msg.add_reaction('✅')
                 await msg.add_reaction('❌')
                 await asyncio.sleep(0.5)
                 try:
                     while 1:
-                        reaction = await self.bot.wait_for('reaction_add', timeout=60000)
+                        reaction = await self.bot.wait_for('reaction_add', timeout=60)
                         if reaction[0].message.id == msg.id:
                             if str(reaction[0].emoji) == '❌':
                                 await msg.delete()
@@ -119,17 +124,22 @@ class CoC(commands.Cog):
                                                           description=f'This dataset is from **{result["season"]}**',
                                                           timestamp=ctx.message.created_at,
                                                           color=DEFAULT_COLOR)
-                                    embed.add_field(name='Tag',value=tag)
-                                    embed.add_field(name='Name',value=result["name"])
-                                    embed.add_field(name='XP',value=result["expLevel"])
+                                    embed.add_field(name='Tag', value=tag)
+                                    embed.add_field(
+                                        name='Name', value=result["name"])
+                                    embed.add_field(
+                                        name='XP', value=result["expLevel"])
                                     # embed.add_field(name='Trophies', value=result["data"][3])
                                     # embed.add_field(name='Attack wins',value=result["data"][4])
                                     # embed.add_field(name='Defense wins',value=result["data"][5])
                                     # embed.add_field(name='Global rank',value=result["data"][6])
-                                    embed.add_field(name='Clantag',value=result["clantag"])
-                                    embed.add_field(name='Clanname',value=result["clanname"])
+                                    embed.add_field(
+                                        name='Clantag', value=result["clantag"])
+                                    embed.add_field(
+                                        name='Clanname', value=result["clanname"])
                                     #embed.add_field(name='Clan badge', value=f'[Link]({result["data"][9]})')
-                                    embed.set_footer(text=f'Requested by n?gga {ctx.author.name}')
+                                    embed.set_footer(
+                                        text=f'Requested by {ctx.author.name}')
                                     await ctx.send(embed=embed)
                                 return
                 except asyncio.TimeoutError:
